Put type code handling code in a separate file
